Custom sparkle slash for stopped analysis
This commit adds a custom sparkle slash symbol image to the status toolbar when the analysis is stopped. The previous red sparkle system symbol was not clear enough to represent the stopped analysis status. To make it more intuitive, I replaced it with the custom sparkle slash image. Additionally, I refactored the BoardView by moving the condition of showing the AnalysisView into the AnalysisView itself. This ensures that the animation of the new symbol image only affects the analysis view and the status toolbar, without affecting the whole board.